feat: move remove_user_from_account function to accounts/utils.py
