+++ helpers/views.py
@@ -1,5 +1,4 @@
 import json
-from typing import Union
 
 from django.contrib.auth.models import User
 from django.shortcuts import render, redirect, get_object_or_404
@@ -139,24 +138,6 @@ def determine_user_role(user: User) -> str:
         return 'is_member'
 
     return 'default'
-
-
-def remove_user_from_account(user: User, account: Union[Community, Institution]) -> None:
-    """Removes user from account
-
-    Args:
-        user: The user object.
-        account: A community or institution account.
-
-    Returns:
-        None
-    """
-    if user in account.admins.all():
-        account.admins.remove(user)
-    if user in account.editors.all():
-        account.editors.remove(user)
-    if user in account.viewers.all():
-        account.viewers.remove(user)
 
 
 def remove_user_from_affiliated_communities_and_institutions(user: User, affiliation: UserAffiliation) -> None:
